refactor(basket): Phase 2 layout move Domain + Contracts
- Move CustomerBasket, BasketProduct to Domain/; IBasketStore to Domain/Abstractions/.
- Move OrderCreatedEvent, ProductPriceUpdatedEvent to Contracts/Integration/.
- Pure namespace rename + using updates; no behavior change.
