+++ basket-microservice/Basket.Service/Domain/BasketProduct.cs
@@ -1,3 +1,3 @@
-namespace Basket.Service.Models;
+namespace Basket.Service.Domain;
 
 internal record BasketProduct(string ProductId, string ProductName, decimal ProductPrice, int Quantity = 1);
