+++ basket-microservice/Basket.Service/Endpoints/BasketApiEndpoints.cs
@@ -1,6 +1,6 @@
 using Basket.Service.ApiModels;
-using Basket.Service.Infrastructure.Data;
-using Basket.Service.Models;
+using Basket.Service.Domain;
+using Basket.Service.Domain.Abstractions;
 using ECommerce.Shared.Observability.Metrics;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Caching.Distributed;
