+++ basket-microservice/Basket.Service/Infrastructure/Data/InMemoryBasketStore.cs
@@ -1,4 +1,5 @@
-using Basket.Service.Models;
+using Basket.Service.Domain;
+using Basket.Service.Domain.Abstractions;
 
 namespace Basket.Service.Infrastructure.Data;
