+++ basket-microservice/Basket.Tests/Domain/CustomerBasketTests.cs
@@ -1,4 +1,4 @@
-using Basket.Service.Models;
+using Basket.Service.Domain;
 
 namespace Basket.Tests.Domain;
